Adjusted loop logic.
It wasn't correctly looping as previously spelled, so now it has been
adjusted to appropriately try a certain number of times before marking
the task a failure.  Previously it would loop infinitely.

In case it seems unclear, if --retry-count is set to 1, then the script
will try *twice* as it is *retrying* the failure.  To make sure the
script does not retry (not sure why one would want to do that)
--retry-count should be set to 0.