Fixed failed host tracking.
Since it was just a list, after we looped over we could continue to keep
adding the same host to the list.  Eventually that's going to exceed 50%
of the total hosts if the number of retries is high enough.  Instead,
re-initialize that list for every loop.

It is safe to reset the list on every loop since we can be generous
here.  If a host starts responding, then that's a good thing!